The build system works now
With the minor difficulty that it appears to be impossible to use a bash
shebang from pwsh-land, even though bash is in the active path. I might
figure this out yet, but in the meantime, I can just launch a bash shell
to generate source lists when I add C++ files.

Also, rearranged the order that some things happen in the cmake
generation. We now compile rust and then attempt (and fail) with the
C++. First steps!