Re-implement settings as ini-backed
The MCM mod itself manages the INI that backs reader settings.
Our job is to read from it. Still missing is any concept of updating
it after the fact. (tl;dr the OnceCell will go away.)

Did a fit and finish pass over the settings themselves and added
the ones I think are necessary. We have 10 settings now, of which
six are required keybinds. I hope to keep it to this number.

Copied in a pile of graphical resources in preparation for making
a run at adapting the UI rendering code. I will leave that in C++,
but I want to read it and edit it as I understand it.

The Rust compiles but the C++ still doesn't.