IT'S ALIVE!
It compiles, links, and builds pdb files.

The notification function was missing a `const`. Ha ha.
Anyway, put it to work responding to menu events, even
though this is not the right way to do it. It works for
now. Also, starting exposing some more CycleEntry stuff
for use elsewhere, since that is what the type is for.