equip-delay timers and their sequelae
We now start equip-delay timers when the player taps a cycle hotkey.
We give them a brief window to keep tapping without us doing anything
other than updating what's visible in the HUD. When the delay expires,
the item is equipped or readied.

The timers are implemented as tick countdowns in the UI renderer.

Changed a lot in the player, gear, magic, and utility_items namespaces
to support this work. My intention is to make the `player` namespace
the entry point for equipping and unequipping. These functions then
call into nitty-gritty implementations in the `equip` namespace. These
are currently split up among the three additional namespaces
mentioned. I moved them around a bit to make the names less of a lie.

The rust side compiles, but as usual I have no idea about the C++
side. Time to find out how many mistakes I made.