Wired up inventory count change hooks
Inventory counts are now updated for utility items and ammo.

The left hand equip logic is working EXCEPT that the player
character doesn't display the item mesh. The item is marked as
equipped, and the right animations play, but no item in hand.
The mesh is fine, because switching to the item by other means
works as expected. So there's something wrong with calling equip
item from an equip-changed hook. Fair enough. Time to figure out
something else.

While I was doing this, renamed `CycleEntry` to `TesItemData`,
since its use is increasingly more as a substitute for passing
TESForm around. I'm not passing TESForm around because the bridge
generator doesn't deal with char*, and tbh I don't want to deal with
C strings either, so I make a container with a helpful name.