More dead code cleanup. Potion logging.
I'm not sure why, but the game is locking up on consuming potions
via the activate key. The same potions are drinkable normally from
the inventory menu, so it's my bug. BUT the consume potion code is
the same (simplified because features are removed) as the previous mod.

I dunno.