Rewrote the entirety of the HUD update logic.
It's now mostly controlled by incoming equip-change events.
The big `update_hud()` function (née `update_equipped()`) is
now very stupid-- it just updates views without trying to do
anything with the information. It is now only called once, at
load time.

The function `handle_item_equipped()` used to be a stub. It's now
implemented, and also it's called `handle_equipment_changed()`.
This logic looks at an incoming item and compares it to what the
game thinks the player has equipped. Here's what we know:

- all events trail the game changes
- de-equip events precede equip events
- we only want to react to equip events

So we compare the equipped item (game source of truth) to the
incoming event item. If they're the same, it's an item-equipped
event, and we update the HUD.