Fixed the MCM menu options, mostly
The problem with them was the papyrus script: I had overrides for
functions that did nothing, when they should have called parent or
not been present at all. I blame copy-pasting before understanding.
One of the classic blunders, like going against a Sicilian when
death is on the line.

The remaining problem is that something is wrong with my translation
files, so none of the localized strings are being loaded. As a
temporary workaround, I've put the text directly in the file. I'll
need to fix this before release.