Made consuming potions & poisons work consistently
Many changes work best when queued as SKSE tasks instead of
done directly in-line. This includes changing counts in the
player's inventory as well as equipping items.

Also observed that we're getting a delta as a signed integer from
inventory-count-changed events, not an absolute count value.
Switched the rust function signatures to i32. While I was at it,
removed as many gear-changes to usize from u32 as I could. I think
it's clearer if we stay 32-bit if the C++ side uses that.