Added a global scale feature for HUD layouts.
This is a new top-level field on the HudLayout struct. To make
this change non-breaking, it's marked as default for serde. If
the renderer sees a zero value here, it substitutes 1.0.

All element sizes, font sizes, and offsets get multiplied by this
scaling factor. This allows you to make a layout at a high
resolution then shrink it down casually, as with a shrink-ray.

Also, added some safeguards for layouts by pinning them to
screen edges if they're positioned in places not within the
current screen size.