Look, I'm just saying overloading was a mistake.
Inheritance was a mistake, yes, but IMO function overloading was
a worse one. And of course operator overloading damns a language
to the lowest circle of hell, but we knew that.

Anyway, I failed to pass the equip slot to the equip-magic
call, which meant it defaulted to the left hand. Surprising, tbh.