Fix #2 : funky behavior with fade-out
Who likes a single source of truth? Programmers, that's who.

Moved the was-visible state variable to the Rust side's persistent
data, and funneled all checks to the controller.

Made a C++ helper function be the single check for whether we want
to draw the HUD or not. This consults game state, the fade-out
setting and player state, and finally the state var that holds
whether the player last requested the hud visibility be.

Cleaned up a whole lot of functions I don't need any more. This
feels better.