Fixed some small problems with H2H
Because it's a not-equipped state, going from a two-hander in the right
to a one-hander does not send the same set of events, so we have to
detect this case differently. Fortunately for me, I had left a debug
line in the right place saying that that spot was where to consider
changing, so I handle the case there. We also need to update the cached
left-hand item when switching to unarmed in the first place.