A complete rewrite of the advancing-cycles codepath (#11)
Plus a significant rework of the timer-expired and equip-changed
functions as well. The logic is now as explicit as I can make it, for
both easy reading now  and easy reading after months away. It's a
little repetitious as a result, but this is fine. To support this, I
added a pile of little types for managing hotkeys and key up/down
events more rigidly. This lets me do more matching instead of
if-equalling. This rework fixes #7 and fixes #8 .

The rewrite allows the left hand cycle button to move away from
two-hander-- it's no longer blocked when you have a bow equipped, say.
This is a new feature.

Made the utility slot accept arrows. Made the HUD update its arrow
count when they're fired. (tl;dr `update_hud()` should be called
_only_ when initializing, not when things change later.)

Added an optional modifier key for the activate shortcut.

In the background, renamed classes and functions to match how they
ended up being used. This adds to the diffs, but is meaningless. There
were no load-bearing changes on the C++ side at all.