You can now cycle length-1 cycles all you want.
Also, better handled the case for powers of the equipped item
not being in your powers cycle. If we detect this, the cycle key
moves to the top of the cycle instead of to the next item. It's
a very cheap bounce-back.

QOL: play the potion consumption sound when you use a potion.
It's also playing the poison sound somehow, I think, so I need
to debug this before I ship.