Probably fix #5: fonts for locales
This turns out to be boolean options that control whether
imgui builds glyphs for the codepoint ranges of interest to
those locales. Added booleans to the config for each toggle-able
range, and defaulted simplified Chinese and Cyrillic to true.
This might have memory-use implications I don't understand yet,
so I'll test to be sure.