Re-implemented fading in & out, with easing.
I had accidentally removed all triggers for fading when I reworked
how visibility is determined. I've decided to be happy about that,
because I was pretty unhappy with how fading looked.

Some notes about the implementation itself:

I'm using a cubic easing function to scale the transition from the
starting point to the end goal. This is subtle but noticeable
in-game. Fading *in* is twice the speed as fading out, on the
theory that seeing the HUD when you need it is more important than
having it go away. We also handle changing fade direction midway
nicely, so the alpha shouldn't jump.

There's also 750 milliseconds of hysteresis on any decision to
auto-fade. The game will give us a brief moment of thinking the
player's weapons are not drawn when changing what's equipped.
We don't want to start fading out every time that happens.

One last note: imgui doesn't support alphas on the full window, only
on individual elements. So any element draw has to remember to
multiply all its color alphas with the global hud alpha.

Did some misc formatting & tidying. Renamed functions so their
usages are more clear to me, a bear of very little brain. Shuffled
some logging around.