FEAT: slow-motion cycling while in combat; long presses; settings rework (#22)
You can now enable slowed time while cycling if you're in combat. This
gives you more time to think about what you're doing when you're in a
tight spot.

Implemented text alignment for layouts. You can now align text for
slots individually, left, centered, or right. Updated the layouts to
use this feature.

Added the _Inter_ typeface to a new layout, the `i18n` layout. (This
is short for "internationalization.") This typeface has support for
many character sets. Also added it to the new `centered` layout.

Added long press as a way to activate things. Reworked the settings
extensively to include this as an option where it made sense.

New settable options:

- How to add an item in a menu: tap, long press, or +modifier.
- How to use consumables: dedicated hotkey, long press, or modifier.
- How to unequip: long press, modifier, add to cycles, or no support.
- How long a long press is.
- How long fading out takes.
- If slow-time while cycling in combat is enabled.
- How much to slow down time if so.

Relaxed the version of MCM Helper we asked for, from 13 down to 11.