Cycles are now stored in the SKSE cosave file (#26)
We now write cycle data to the SKSE cosave file on game save, and read
it on game load. This should drastically improve the stability of
cycle data for players, since it will stay with a single character as
they switch, or if they rename a character.

Some implementation notes:

The TOML serialization is still there as a backup, and indeed is
required on the first load of a cosave-supporting plugin.
Deserialization will need to be present until at least a 1.0 release,
as removing it would be a breaking change. 

SKSE cosave serialization hands you a byte writer and reader and lets
you define your own serialization format. So we take cycle data,
serialize it into a bag of bytes, then write the entire bag of bytes.
I'm using [rkyv](https://lib.rs/crates/rkyv) as a reasonable time/size
tradeoff in the serialization format. My cycles are around 1460-ish
bytes in length, which is just fine.