Code cleanup & rearrangement
Made a facade file for all the functions the controller exposes
to C++. The controller file is still too long for comfort.

Make things temporarily worse by adding another enum. This one is
`CycleSlot`. It's not shared, so it has nice behavior. It cleans up
the cycle implementation while making the controller implementation
slightly messier. It also allows me to push the Action enum out
toward the boundary a little bit.