FEAT: You can optionally link favorites to cycles
This is a new setting that connects favorites with cycles.
If you favorite an item, it's added to the appropriate cycle
with our best guess about hand for weapons.

If it's a two-handed weapon, it goes in the right hand.
If you have two or more of the weapon, it goes in both.
Spells go in both. Powers go into the powers slot, etc.

Unfavoriting removes the item from the same slot it went into.

The things I learned from this exercise allowed me to send more
information to the handle-equipped hook, which make the logic
there more reliable. I also know how to figure out if a weapon
is poisoned, and how to get enchantment levels.

This is actually pretty cool.