An optional new icon set
I found a perfect set of icons on the Noun Project and paid to use
them. Extended spell and armor classification enough to put some of the
new icons to use in new icon kinds. The integration is not complete,
but it's complete enough to test my work figuring out spell and armor
subtypes. There are fallbacks for layouts without the additional new icons.

Made a layout that includes these icons and tweaks slot sizes a bit.
Added a just recipe to build the new layout archive.