Handle strings that aren't valid UTF-8 without crashing
Rust strings are always valid UTF-8. Windows-native strings are not;
apparently they're not even necessarily valid UTF-16. There are two
places in the mod code where this might be an issue, and players have
reported crashers related to both.

First case: logging to a file path that itself is not valid UTF-8, as
well as saving the TOML version of the cycle data with a player name
that is not valid UTF-8. This is what OsString was designed to cope
with. PathBuf assumes utf-8, but the `std::file` api accepts OsString.
So for these cases, we get a Vec<16> from the C++ side and then
construct an OsString from it.

The second case is item names, which might come from mods that are
free to use whatever characters they wish. An item name, like the
player name, is returned from the `TESForm` item as a C string. It
uses characters that the Flash-based menus can display, which
appear to want UCS-2 aka windows wide strings, sort of UTF-16 BE,
which is why the translation files require that encoding. AFAICT. In
this case, what we need to do is use some kind of encoding conversion
at display time to re-encode the characters. I am not doing that.
Instead I am doing this:

We now extract the bytes into a Vec<u8> and pass that over to the Rust
`ItemData` struct. This struct attempts to make a valid utf string out
of it. If it can, that's what it uses to show the data. If it can't,
it uses a lossy string. The same bytes are handed to `imgui`'s text
drawing API as a C string, which does the same lossy display.

This is sub-optimal, but far better than crashing.

I have tested the item name case to my satisfaction. I have not yet
tested the filepath case.

This PR should fix #32 "handle non-UTF-8 data".