Feature: potions can be grouped & auto-selected
You can now optionally group magicka, health, and stamina potions
together into a single entry in the utilities cycle. If you consume
one of these proxy entries, the HUD selects the potion that comes
closest to refilling your stat deficit without overflowing. If you're
not down on the stat, the HUD will honk (very quietly) and decline to
consume a potion.

The totals for these grouped potions in the HUD are the totals of all
of that potion kind in the player's inventory. While I was making the
HUD update the counts for these, made it update counts for everything
else at the same time.

There is an MCM toggle for the feature and matching translation
strings.

Made a number of improvements to how and when the HUD updates what
it's showing.

Improve equipped item sink reliability by no longer reling on extra
data, but instead asking directly for what's in the left and right
hands and comparing.