Cosaves can be versioned; move to a version 1
Version 1 of the cosave format is a list of form strings with
no additional data, so we can de-serialize into anything.
Including, cough cough, the new struct used by the HUD in the
icon branch.

The Rust side now informs the C++ side which version number to
use when saving. It successfully reads v0 and v1, and writes v1.

Removed all code for writing TOML files to persist cycle state.
We can no longer read cycle data from TOML files, so this is a
breaking pre-1.0 change. It will require a minor version number
bump. We still use TOML for reading layouts, so the deps remain.