Resurrected the data tests
Satisfied clippy at least partially (for now).
Simplified the color enum. Got rid of a lot of now-unnecessary
serialization code (we only serialize form strings now!). Made
all color extraction use the bottleneck.

Most importantly, introduced the concept of a HudItem LRU cache.
The controller manages this, and it can now prefer to use the
cache rather than get every interesting bit of item data from the
C++ side every time the two need to pass item data back and forth.
The form spec string is the cache key, and this is unique for each
TESFormItem we track, so we can now pass that back and forth.

This isn't yet in use everywhere, but it now exists.