OCF keyword support plus a hud item cache (#43)
This PR lands a long-running branch with a lot of loosely-related
work.

Most significantly, the `ItemData` struct and the `ItemKind` enum have
been replaced by HudItem and a number of enum types not shared with
C++. These new types are the core of how we now classify items and
answer questions about them at runtime. They are true Rust sum types,
not C enums, so exhaustive matching is possible. 

The `Icon` enum replaces `ItemKind`. It is invisible to the C++ side.
The C++ side now does only top-level rough TES form categorization.
its job now is to collect the information that the Rust side will use
to select the precise icon and color to use to represent that item in
the HUD. There are three kinds of classification: keyword
classification (armor and weapons), spell and spell effect data
classification (spells, scrolls, shouts), and simpler classification
for straightforward items like powers. 

There's an entire new sub-mod in the `data` directory with types for
game enums like actor values and spell archetypes, then stricter types
for classification as the much simpler HUD representation. Much of
this is also about reading OCF keywords from form info and using those
to decide what icon to use for armor.

To make all this work have a visual effect, this PR includes a new
icon set and a layout for that icon set. The icon set is licensed from
the Noun Project under my name. It is the RPG icon set from MaxIcon
set, and it is amazing. The only downside to it is that it is not the
same as the SkyUI icon set used elsewhere. I am leaning into spell-
and shout-specific icons, however, which I have not seen from the I4
project as much, so maybe people will like it. It is optional.

The cache work was unrelated to the classification work, but enabled
by it. All HudItem creation (or nearly all) is mediated by access to
an LRU cache, keyed by the formspec string. The formspec string is
used to serialize enough info to uniquely identify an item. It takes
the form `Plugin.esp|0xdeadbeef`, recording the originating plugin and
the form id. The cycle data struct is serialized as that string and
only that string, because it's sufficient. The C++ side now sends only
that form string over to Rust to inform it of item-related events, and
the Rust side requests fresh data only if it needs. Counts are updated
in the cached version of the HudItem. 

My current game runs around 22-25 items in the cache, but it'll hold
100 before evicting. I do not think memory pressure is significant
here. At the moment the visibility struct in the controller holds
copies of the cached items, not pointers, but I'll probably make that
indirect as well. This is where I become performance-sensitive,
however, because that is the hot path for drawing the HUD. Now that I
know how to profile DLLs on Windows I can profile this. (I wonder how
the Rust side will get profiled. Hmm.)

Incidental work done along the way: Removed all TOML serialization
code from cycle data. Cleaned up all now-unneeded use of serde derives
through the Rust code. Renamed icons to fit my fussy least->most
specific naming habit. Small bugs nailed while I was play-testing. 

Work that needs to complete before I release a version with this work:

- classifying food and drink at least a little bit
- testing how the classification behaves without OCF in place (that
  is, should I make it a hard or soft dependency?)
- further classification of spells and shouts, and testing of same in
  a game
- more testing of the cache plus profiling
- fixing any bugs revealed
- removing as much ItemKind + ItemData code as possible, leaving only
  enough to support deserializing from older cosaves