Layouts can choose to hide ammo when no ranged equipped
There are two new booleans for layouts consulted if a ranged weapon
is equipped:
    hide_ammo_when_irrelevant
    hide_left_when_irrelevant
(These names are likely to change.)

They are independent booleans, but both are consulted when a ranged
weapon is equipped. Layouts can choose to "swap in" the ammo for the
left hand slot when a bow is equipped, or merely hide the ammo when
a ranged weapon is not equipped.

Updated the default layout to use this feature.