Fix a 1.5.97 crasher with the menu options MCM Helper function
There's a difference in behavior between the version built for 1.5.97
and the version built for 1.6.640. The SE build needs an empty string
array sent in the short options parameter, even though it should default
to an empty array. The AE version behaves as the docs say it should.
Though perhaps the docs have changed too?