Dry-coded being able to set the 4-char "unique" identifier.
I think I have a case of ids colliding, though I can't prove
it. Since there is nothing forcing the ids to be truly unique,
IMO it's a bug in any mod not to allow for changing the id to
avoid collisions. This of course breaks save data continuity
when the ID is changed, but this is better than the other mod
crashing, which is what I think is happening.

The identifier can be set in the INI. If not present, it defaults
to 'SOLS'. There is no UX for changing the identifier, but I'll
document it for users in a pinch.