feature: the SKSE identifier can now be changed in the mod settings (#48)
Each mod that wants to save data in the SKSE cosave must identify
itself with a "unique" four-character identifier. The word unique is
in quotes because nothing enforces uniqueness across the mod
ecosystem, such as a registration system. Instead the identifier is
assumed to be unique, and at runtime SKSE associates cosave data
records tagged with the identifier with whatever plugin registered
itself first with that identifier. Or so I assume; I haven't read all
of this code yet.

A mod trying to read cosave data written by a different mod might or
might not behave correctly. The symptom observed is that no cosave data
is written at all. Why? I don't know. I haven't reproed it, nor have I
seen any logs with the problem.

My theory is that this is what is causing a specific user to suffer
missing  cosaves when SoulsyHUD is added to their mod list. There is no
other touchpoint, and Soulsy has been successfully saving and reading cosave
data for months now. This build of the mod will let that user-- and indeed
any other user with this problem-- change the ID string to something else.
I suggest `BOOP`. Or maybe `BOP!`. Four ascii characters, anyway.
The identifier can be set either by editing a pre-existing
`SoulsyHUD.ini` or visiting the second MCM settings page.

I'm nervous about exposing this in the mod settings, but people who
are having trouble with corrupted/missing cosaves will probably need
to use this to fix their problem. As a wise man notes, given that
uniqueness is not enforced, it's a bug not to allow setting it on the
fly to work around collisions.