Abstracted cycle behavior into traits
There are three: one for shared behavior, one for equippable things,
and one for equipsets. Implemented the traits for vecs of cycle-entry-data,
which is either an equipset or a form-spec string.

Updated cycle data to use these new traits. This cleaned up
a lot of code nicely.

Started implementing equipset features like renaming.

The main unfinished work is:
- the MCM menu implementation
- debugging player::getEquippedItems() to see if the implementation
  I came up with is at all reasonable