Rework how icon fallbacks are implemented
If all else fails, an icon can suggest a fallback for itself.
These fallbacks come from the core set that's bundled with the
mod. If nothing is found for the core icons, the mod installation
is considered to be broken.

Performed a long-deferred refactor of how weapons and armor are
classified so they use the same efficient approach as spells: the
important information is cached so we do not repeat work when we
render the item. Thus do we keep our performance decent.