Cleaned up some logging and some HUD flickering.
The transition delay timer was exhibiting some odd behavior
when it was triggered just after a transition in the other
direction commenced. The fix is to look at the transition
time for going the other direction, and just snapping to the
goal alpha if it's less than a millisecond.

I'm happy to have finally nailed that one.