Guard against being run in other game data contexts
E.g., Enderal. We were assuming in a few places that a specific
"dummy dagger" item existed, as well as a couple of sound objects.
Added some null-proofing for the sounds. The proxy dagger was
being used to force-un-equip hands by equipping and then unequipping.
We still need to do that when unequipping spells, but we can avoid
it when unequipping bound objects. There will be a minor problem
in non-Skyrim game data cases, where the spell effect might remain
visible in a hand that the game thinks is empty, but this is better
than crashing.

This code was the only thing remaining in weapons.{h,cpp} so
I folded it into gear.{h,cpp}.