Fixed text alignment
The text box size calculation uses the size that was specified
when the font was loaded. If that size is way off from the size
it's being rendered at-- say because there's a significant global
scaling factor-- the offsets were all wrong. The fix is to use
the scaling factor when loading the font, and give people a tip
in the layout docs that the global font size hint is meaningful.

Also documented the interpretation of alignment, which is to
change what the offset means as well as how the text is moved.
The offset is interpreted as left top corner when using left
alignment, middle when centering text, and right when
using right alignment. This works with how people think about
where they want to locate the text.