Work around the MCMHelper on 1.5.97 for now
We work around it by hiding the cycle preview menus on the
maintenance page. I can distinguish between builds and change
behavior, but what works for me on pre-AE builds doesn't work
reliably enough for other players.

I think it would be fixed by a rebuild and re-release of the mod,
BUT my local builds of it are not working right for reasons I have
yet to debug. While I work on that, disable the feature for
1.5.97 users. :(