All player messages are now translatable
Added format strings with named arguments to the scaleform
translation files. Look up the strings from the translation files
when we're constructing player-facing messages instead of using
hard-coded English messages.

Messages might fail to be formatted, in which case we say nothing
to the player about what their action accomplished.

I'm not yet doing anything to cope with wide strings from the
scaleform files, which are UCS2 BE files and might contain invalid
utf8. Sigh.