More fun with the MCM.
Moved the tricky preview option to its own page, so that all
settings are available sans potential to crash. Rang yet another
change on the sending-options-to-mcmhelper problem. The trouble
is indeed with indexing into the short names array. Do the strings
need null termination? IDK.

Cleaned up the papyrus functions a bit.