This is the fix for the cycle preview crasher.
I believe I have proven it to my satisfaction. If it is not
a fix, I will do something drastic like eat an entire pint of
ice cream in one sitting. There is no ice cream in my house, so
this would obviously be traumatic.

The root bug is in how MCM-Helper handles empty dynamic menus,
and it's been around for a while. The workaround is to never
give MCM-Helper an empty array, but instead to show the string
"-- empty --" (or its translation) in the drop-down.

This is a PR with a fix for MCM-Helper:
https://github.com/Exit-9B/MCM-Helper/pull/16