Fix #53 : bound weapons aren't showing up properly
Yeah, because the weapon HUD item builder was skipping bound
weapons. Why? I bet this decision was inherited from that
initial fork, but I should have caught it during the rewrite.

Added a special color for bound items so they look all magical.

Added some logging in the hud item constructor to report if
an item that should be shown in the HUD is falling back to an
empty item, to catch cases like this more quickly.