Differentiate torches from lanterns.
Check the OCF formlist (which hasn't included anything for me
yet), then fall back to the name.

This particular solution is hacky and only works for
English, because it looks at the item name. LIGH form
types do not have keywords associated with them. A hacky
detection is better than no detection.