Equip sets can now be created and cycled!
Equipping worked first try! How unlikely is that?
Untested: updating, renaming.

- Provide a parse-string-as-int function for papyrus to use.

- Equip sets can be selected in the MCM and renamed/updated/deleted.

It's not the user experience I want, but everything works.
Menus have to have string values.  Right now I'm using string
versions of equip set IDs as those values. Showing names and returning
string ID values would be better, as would not jumping through hoops
to store an index into the array and then get that back as the set ID
again later. But as they say, first make it right, then make it fast.
Or tidy.

- Add empty slots to equip set data & unequip if requested

Equipment sets also need to track which slots are empty so we
can un-equip items when the player puts on an equipment set.
For example, they might want to remove a cloak or a backpack.

Added a setting to equip sets to unequip empty slots. Added UX
for toggling the setting. Track the empty biped slots, write them
to and read them from cosaves, and unequip if the setting requests it.

- Equipment sets now store icons for drawing in the HUD.

Or more precisely, they now store the name of the item to use
as their icon. While I was poking at icons, cleaned up HasIcon in
a way I've wanted to do for a while, which is to reduce it to
color() and icon() functions. Icons know their own files, and also
their own fallbacks. (Mostly. Spells are an exception.)

There's no way to assign an icon source yet, so we're taking the
first item in the equipset list. That UX is next...

- Icons can now be assigned to equipment sets.

They round-trip through saves. They are drawn in the HUD by
layouts that support equipment sets. Equipping sets works.
The unequip-empty-slots setting works. Added some icon checks
to the tests.

The information dump in the logs at load time now includes
equipment sets.

This feature is very close to shippable.