Improved handling of unequipped items.
Changed the signature of the Rust function that handles
this to allow the C++ side to send more information over
from the jump.

Also, fixed a bug with trying to dual-wield weapons by
not sending extra data along with a weapon equip request.
I am not sure why this works, but it does. (Maybe the equip
manager notes from the extra data that the weapon is already
equipped in the other hand slot and declines to act? IDK.)