fix #59 : fire long-press actions on a timer instead of on key-up
And with this checkin, the work for firing long-press on a timer
is complete. It is in fact nicely snappy. Feels great.

We figure out what the requested action is then use existing
functions to get it done. We then stop any timers associated
with that cycle.

Cleaned up the key-up handlers for right & left hands  a little
bit now that they don't have to handle long presses. This logic
is still in the wrong place overall. It feels like it can be
bundled up more nicely. E.g., the Action enum is really an Event,
and the Event data includes any modifiers pressed. And the settings
determine which, if any, things should happen on that key down or
key up.