Fix a multitude of sins with getting inventory updates
Specifically for items that are picked up. We have to go
to the base object for its form id and look that up. Then,
if we do not have the item in our cache, we must populate
the cache. Then we can reliably update the HUD with count
changes on the fly.

Also, observed that we were getting unequip events for
ammo partial drops, so we double-check the form id against
the currently-equipped ammo's form id to see if it's really
been unequipped or not.

It is nice to see the potion counts change properly!