Fix #66 Hud retains the utility item from the last-loaded save
This was a bug in controller.update_hud(). It was overly-cautious
about updating utility items on game load, which is not a time
to be cautious about updates. While I was there, observed that
this function is only ever called internally by the controller,
so I de-pubbed it and removed it from the C++ bridge.