fix #58 - make the mod work properly with CGO
This was some interesting work! We now subscribe to animation graph
events, so we can get the CGO grip change events. If we see one,
we set a state var on the controller object, and change our HUD
view to show what CGO is showing.

This caused some refactoring to pull out some logic that is now
re-used in several situations where we're switching to
effective-2-hander from effective-1-hander.