Remove all traces of nanosvg from the project
Implemented a rasterize-by-file-path function for svgs,
wrote a test for it, and exposed it to C++. Generalized the inner
rasterization function to scale if given a max dimension size,
otherwise use the identity transformation.

Converted the second texture-loading code path to use the new
Rust function. This was the last use of nanosvg in the project,
so removed the dependency and tidied up cmake.

There's more cleanup to do, but it's time to test.