Suppress all inventory events we don't care about
We make a set for the enum variants for form types
that might possibly be shown in the hud, and filter out
events for anything that isn't one of those.

Lots of logging tweaks to clean up debug-level logging
for human consumption. The gnarly stuff is now at trace
level.