Fix #69 - crashes in the favorites menu (nice)
I rewrote a chunk of the code that determines what the selection
was in a menu when the player presses a key we care about.
Specifically, we now look at the regular form item and do not
care about bound items unless we have to. Spells and shouts don't
have them. This eliminates one category of crash.

The other fix is that we no longer pay any attention at all
to the toggle-favorites hotkey when in the favorites menu itself.
We stil pay attention to cycle hotkeys.

I hate the menus and how inconsistent they are. FYI. HTH.