Menu key events now not passed on if we handled them
This fixes adding and removing items in menus for controller
users. In fact, if you want to, you can just straight up use
the four dpad keys without modifiers. They will no longer work
for navigation in the menus, but the thumbsticks *will*.

If we decided to act on a key pressed in the menu, we suppress
it so nobody else acts on it. We don't bother seeing if it overlaps,
because we don't care. We acted so nobody else should.

NOTE: This required changing from handling menu events on key
up to handling them on key down, so we can beat the damn flash
app to the punch. This does change the feel a little bit.

While I was testing this, noticed and fixed a minor bug in the MCM
with showing the long press to match hands option when the
unequipping option selection changes to allow it.

Because I couldn't help myself, did a little more renaming for
style guide compliance. You know how strict ceejbot is about style
guide compliance. The C++ side of this codebase is in a constant
state of INFRACTION.