Better reporting for layout parse failures
We use a little bit of eyre's reporting features now.
Also, on parse failure we try parsing with each layout version
to get detailed errors out of serde instead of just the error
about none of the untagged enum variants being matched.

For extra points, we display a message one the screen when
the layout load fails. This message is translatable.