Icon names are now generated from enum variant name
This means the naming convention is now enforced. Renamed the
icons that didn't fit the convention. Separated the THICC icon
checking test from the Soulsy icon checking test, and marked it
as ignored so we don't fail test runs. (There are a lot of
icons missing from that pack.)

Removed the annoying Windows execute bit from some icons.