Fix #76 : Handle the Ordinator "Boundless Cup" perk
The way we handle this perk is to apply poisons completely
differently than we had been handling them. The poison code
was a holdover from the origins of this mod in Lamas Tiny Hud.
That mod did all the work itself-- calculated doses and
applied poisons to both left and right hand weapons using
the `PoisonObject()` function on the weapons.

This implementation preserves the left & right hand feature
while going through the game's perk application flow instead of
needing to replicate it. We use `EquipObject()`, passing the
player actor, the poison, and the slot to equip it to, with
either or both of the left and right hand slots. (Assuming the
player has enough items.)

Simpler and more robust ftw.