Dry-coded the poison indicator feature.
There's a new optional per-slot element in layouts v2 (electric
boogaloo). This element displays a poison indicator if the item in
the slot is poisoned. It's only meaningful for left/right hand slots,
but we make no effort to enforce this; hud items in those slots will
never say they're poisoned.

The indicator looks like this:

```toml
[left.poison]
offset = { x = 0.0, y = 50.0 }
[left.poison.indicator]
svg = "icons/indicator_poison.svg"
size = { x = 23.0, y = 23.0 }
color = { r = 255, g = 255, b = 255, a = 255 }
```

HUD items have a function `is_poisoned()` for checking this bit.
This call goes back to C++ to look at the associated bound object's
extra data to see if it's poisoned. We are checking this on every
draw loop right now. I'm not sure caching it would be a win, because
I have not yet investigated whether there's any reasonable invalidation
notification.

The renderer renders this layout element if the item is poisoned
and if the color for the image has non-zero alpha.

This feature is only supported on layouts v2.

This addresses the poison indicator part of bug #38.